+++ src/parser/sketch/components/AbstractText.js
@@ -5,5 +5,6 @@ initAbstractComponent();
 export default class AbstractText extends AbstractComponent {
     constructor(data){
         super(data);
+        // console.log(data);
     }
 }
